First commit of a "MiniFWLite" library whose sole purpose is to suppress missing dictionary warning messages when using our ntuples in pure ROOT. Most of the messages are completely spurious, remnants of the fast-clone process.  For these classes of which no actual data members reside in the ntuple there are empty class definitions just to keep ROOT quiet.
The only legitimate warnings are for the edm::Wrapper<T>'s.  For this I
stripped more than 90% of the junk we don't need from the Wrapper class
and built new dictionaries that just preserve the data members and a
couple methods.  The effect of using this library is strictly cosmetic.
If it is ever found to be otherwise then may the proverbial god have mercy
on all of our proverbial souls.